Add Alt property to GravatarImgHelper and clean up code
Added a new property `Alt` to the `GravatarImgHelper` class with a
default value of "Gravatar image". Modified the `Process` method to
use the new `Alt` property for the `alt` attribute of the `img` tag
instead of the hardcoded "Gravatar image". Removed comments and
redundant lines in the `GetMd5Hash` method to clean up the code.